variuos improvements
