+++ src/main/java/com/github/skjolber/unzip/ChunkSplitterFileEntryHandler.java
@@ -1,0 +1,108 @@
+package com.github.skjolber.unzip;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.InputStream;
+import java.util.concurrent.ThreadPoolExecutor;
+
+/**
+ * Read and split inputs into larger parts based on newline. 
+ * 
+ */
+
+public class ChunkSplitterFileEntryHandler implements FileEntryHandler {
+
+	protected final int chuckLength; // effective length depends on line lengths
+	protected final ChunkedFileEntryHandler delegate;
+	
+	/**
+	 * Constructor.
+	 * 
+	 * @param chuckLength number of bytes per segment
+	 * @param delegate delegate for forwarding (whole or partial) streams. 
+	 */
+	
+	public ChunkSplitterFileEntryHandler(int chuckLength, ChunkedFileEntryHandler delegate) {
+		super();
+		this.chuckLength = chuckLength;
+		this.delegate = delegate;
+	}
+
+	public void handle(final String name, long size, InputStream in, final ThreadPoolExecutor executor, boolean consume) throws Exception {
+		if(size > chuckLength) {
+			
+			FileChunkSplitter splitFileEntry = delegate.splitFileEntry(name, size);
+			if(splitFileEntry != null) {
+				byte[] buffer = new byte[Math.min(8192 * 16, chuckLength)];
+	
+				ByteArrayOutputStream bout = new ByteArrayOutputStream(chuckLength);
+				
+				while(true) {
+					
+					int remaining = chuckLength;
+					
+					int read;
+					do {
+						read = in.read(buffer, 0, Math.min(buffer.length, remaining));
+	
+						if(read == -1) {
+							break;
+						}
+						
+						remaining -= read;
+	
+						bout.write(buffer, 0, read);
+					} while(remaining > 0);
+					
+					final byte[] byteArray = bout.toByteArray();
+					
+					if(read == -1) {
+						// end of file
+						delegate.handle(name, byteArray.length, new ByteArrayInputStream(byteArray), executor, false);
+						
+						break;
+					} else {
+						int index = splitFileEntry.getChunkSplitIndex(byteArray, byteArray.length - 1);
+	
+						if(index == -1) {
+							throw new IllegalArgumentException("No newline found in chunk size " + byteArray.length);
+						}
+						delegate.handle(name, size, new ByteArrayInputStream(byteArray, 0, index - 1), executor, false);
+	
+						// reuse buffer
+						bout.reset();
+						// write tail
+						if(index + 1 < byteArray.length) {
+							bout.write(byteArray, index + 1, byteArray.length - index - 1);
+						}
+					}
+				}
+			} else {
+				delegate.handle(name, size, in, executor, consume);
+			}
+		} else {
+			delegate.handle(name, size, in, executor, consume);
+		}
+	}
+
+	@Override
+	public void beginFileEntry(String name) {
+		delegate.beginFileEntry(name);
+	}
+
+	@Override
+	public void endFileEntry(String name, ThreadPoolExecutor executor) {
+		delegate.endFileEntry(name, executor);
+	}
+
+	@Override
+	public void beginFileCollection(String name) {
+		delegate.beginFileCollection(name);
+	}
+
+	@Override
+	public void endFileCollection(String name, ThreadPoolExecutor executor) {
+		delegate.endFileCollection(name, executor);
+	}
+	
+}
