+++ src/main/java/com/github/skjolber/unzip/ChunkedFileEntryHandler.java
@@ -1,0 +1,7 @@
+package com.github.skjolber.unzip;
+
+public interface ChunkedFileEntryHandler extends FileEntryHandler {
+
+	FileChunkSplitter splitFileEntry(final String name, long size);
+	
+}
