+++ src/main/java/com/github/skjolber/unzip/FileChunkSplitter.java
@@ -1,0 +1,11 @@
+package com.github.skjolber.unzip;
+
+/**
+ * Chunk splitter - determines chunk boundary
+ * 
+ */
+
+public interface FileChunkSplitter {
+
+	int getChunkSplitIndex(byte[] bytes, int fromIndex);
+}
