--- src/main/java/com/github/skjolber/unzip/NewLineSplitterFileEntryHandler.java
@@ -1,7 +1,0 @@
-package com.github.skjolber.unzip;
-
-public interface NewLineSplitterFileEntryHandler extends FileEntryHandler {
-
-	boolean splitFileEntry(final String name, long size);
-	
-}
