+++ src/main/java/com/github/skjolber/unzip/csv/CsvLineHandler.java
@@ -1,13 +1,12 @@
 package com.github.skjolber.unzip.csv;
 
-import java.util.Map;
-
-public interface CsvLineHandler {
+public interface CsvLineHandler<T> {
 
 	/**
-	 * Handle line in key-value form
-	 * @param fields key-value map, which will be reset by the caller before reading the next line
+	 * Handle line in key-value form.
+	 * 
+	 * @param value line value
 	 */
-	void handleLine(Map<String, String> fields);
+	void handleLine(T value);
 	
 }
