+++ src/test/java/com/github/skjolber/unzip/ZipFileTest.java
@@ -11,7 +11,7 @@ public class ZipFileTest {
 
 	@Test
 	public void testMultiThread() throws IOException {
-		FileEntryHandler handler = new NewLineSplitterEntryHandler(32 * 1024 * 1024, new TestCsvFileEntryHandler(new NoopCsvLineHandlerFactory()));
+		FileEntryHandler handler = new ChunkSplitterFileEntryHandler(32 * 1024 * 1024, new TestCsvFileEntryHandler(new NoopCsvLineHandlerFactory()));
 		
 		File file = new File("./src/test/resources/static/rb_norway-aggregated-gtfs.zip");
